[Tests] Tests for json payload
