+++ engine/flink/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/confluent/ConfluentSchemaRegistryProvider.scala
@@ -7,6 +7,7 @@ import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies
 import pl.touk.nussknacker.engine.avro.schemaregistry.{SchemaRegistryError, SchemaRegistryProvider, SchemaRegistryUnsupportedTypeError}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.{CachedConfluentSchemaRegistryClientFactory, ConfluentSchemaRegistryClient, ConfluentSchemaRegistryClientFactory}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.formatter.ConfluentAvroToJsonFormatter
+import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.jsonpayload.{ConfluentJsonPayloadDeserializerFactory, ConfluentJsonPayloadSerializerFactory}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.{ConfluentAvroSerializationSchemaFactory, ConfluentKafkaAvroDeserializationSchemaFactory}
 import pl.touk.nussknacker.engine.avro.serialization.{KafkaAvroDeserializationSchemaFactory, KafkaAvroSerializationSchemaFactory}
 import pl.touk.nussknacker.engine.kafka.{KafkaConfig, RecordFormatter}
@@ -39,21 +40,45 @@ object ConfluentSchemaRegistryProvider extends Serializable {
     ConfluentSchemaRegistryProvider(CachedConfluentSchemaRegistryClientFactory(), processObjectDependencies)
 
   def apply(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory, processObjectDependencies: ProcessObjectDependencies): ConfluentSchemaRegistryProvider =
-    ConfluentSchemaRegistryProvider(
+    avroPayload(
       schemaRegistryClientFactory,
       processObjectDependencies,
       formatKey = false
     )
 
   def apply(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory,
+            serializationSchemaFactory: KafkaAvroSerializationSchemaFactory,
+            deserializationSchemaFactory: KafkaAvroDeserializationSchemaFactory,
             processObjectDependencies: ProcessObjectDependencies,
             formatKey: Boolean): ConfluentSchemaRegistryProvider = {
     val kafkaConfig = KafkaConfig.parseConfig(processObjectDependencies.config)
     new ConfluentSchemaRegistryProvider(
       schemaRegistryClientFactory,
-      new ConfluentAvroSerializationSchemaFactory(schemaRegistryClientFactory),
-      new ConfluentKafkaAvroDeserializationSchemaFactory(schemaRegistryClientFactory),
+      serializationSchemaFactory,
+      deserializationSchemaFactory,
       kafkaConfig,
       formatKey)
   }
+
+  def jsonPayload(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory,
+              processObjectDependencies: ProcessObjectDependencies,
+              formatKey: Boolean): ConfluentSchemaRegistryProvider = ConfluentSchemaRegistryProvider(schemaRegistryClientFactory,
+    new ConfluentJsonPayloadSerializerFactory(schemaRegistryClientFactory),
+    new ConfluentJsonPayloadDeserializerFactory(schemaRegistryClientFactory),
+    processObjectDependencies,
+    formatKey)
+
+
+
+  def avroPayload(schemaRegistryClientFactory: ConfluentSchemaRegistryClientFactory,
+            processObjectDependencies: ProcessObjectDependencies,
+            formatKey: Boolean): ConfluentSchemaRegistryProvider = {
+    ConfluentSchemaRegistryProvider(
+      schemaRegistryClientFactory,
+      new ConfluentAvroSerializationSchemaFactory(schemaRegistryClientFactory),
+      new ConfluentKafkaAvroDeserializationSchemaFactory(schemaRegistryClientFactory),
+      processObjectDependencies,
+      formatKey)
+  }
+
 }
