+++ engine/flink/avro-util/src/main/scala/pl/touk/nussknacker/engine/avro/schemaregistry/confluent/serialization/AbstractConfluentKafkaAvroSerializer.scala
@@ -1,14 +1,13 @@
 package pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization
 
-import java.io.{ByteArrayOutputStream, IOException}
+import java.io.{ByteArrayOutputStream, IOException, OutputStream}
 import java.nio.ByteBuffer
-
 import io.confluent.kafka.schemaregistry.avro.{AvroSchema, AvroSchemaUtils}
 import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException
 import io.confluent.kafka.serializers.{AbstractKafkaAvroSerializer, AbstractKafkaSchemaSerDe}
 import org.apache.avro.Schema
 import org.apache.avro.generic.GenericContainer
-import org.apache.avro.io.EncoderFactory
+import org.apache.avro.io.{Encoder, EncoderFactory}
 import org.apache.kafka.common.errors.SerializationException
 import pl.touk.nussknacker.engine.avro.schema.{AvroSchemaEvolution, DatumReaderWriterMixin}
 
@@ -28,40 +27,45 @@ class AbstractConfluentKafkaAvroSerializer(avroSchemaEvolution: AvroSchemaEvolut
     if (data == null) {
       null
     } else {
-      val avroSchema = avroSchemaOpt.getOrElse(new AvroSchema(AvroSchemaUtils.getSchema(data, this.useSchemaReflection)))
+
+      val (avroSchema, record) = (avroSchemaOpt, data) match {
+        case (Some(schema), record: GenericContainer) => (schema, avroSchemaEvolution.alignRecordToSchema(record, schema.rawSchema()))
+        case (Some(schema), other) => (schema, other)
+        case (None, other) => (new AvroSchema(AvroSchemaUtils.getSchema(data, this.useSchemaReflection)), other)
+      }
+
       try {
-        val subject = getSubjectName(topic, isKey, data, avroSchema)
-        val schemaId = if (this.autoRegisterSchema) {
-          this.schemaRegistry.register(subject, avroSchema)
-        } else {
-          this.schemaRegistry.getId(subject, avroSchema)
-        }
-
-        val record = data match {
-          //We try to convert record to provided schema if it's possible
-          case record: GenericContainer if avroSchemaOpt.isDefined => avroSchemaEvolution.alignRecordToSchema(record, avroSchema.rawSchema())
-          case _ => data
-        }
-
+        val schemaId: Int = autoRegisterSchemaIfNeeded(topic, data, isKey, avroSchema)
         writeData(record, avroSchema.rawSchema(), schemaId)
       } catch {
-        case exc: RestClientException =>
-          throw new SerializationException(s"Error registering/retrieving Avro schema: " + avroSchema.rawSchema(), exc)
         case exc@(_: RuntimeException | _: IOException) =>
           throw new SerializationException("Error serializing Avro message", exc)
       }
     }
   }
 
+  private def autoRegisterSchemaIfNeeded(topic: String, data: Any, isKey: Boolean, avroSchema: AvroSchema) = {
+    try {
+      val subject = getSubjectName(topic, isKey, data, avroSchema)
+      if (this.autoRegisterSchema) {
+        this.schemaRegistry.register(subject, avroSchema)
+      } else {
+        this.schemaRegistry.getId(subject, avroSchema)
+      }
+    } catch {
+      case exc: RestClientException =>
+                throw new SerializationException(s"Error registering/retrieving Avro schema: " + avroSchema.rawSchema(), exc)
+    }
+  }
+
   protected def writeData(data: Any, avroSchema: Schema, schemaId: Int): Array[Byte] = {
     val out = new ByteArrayOutputStream
-    out.write(AbstractKafkaSchemaSerDe.MAGIC_BYTE)
-    out.write(ByteBuffer.allocate(AbstractKafkaSchemaSerDe.idSize).putInt(schemaId).array)
+    writeHeader(data, avroSchema, schemaId, out)
 
     data match {
       case array: Array[Byte] => out.write(array)
       case _ =>
-        val encoder = this.encoderFactory.directBinaryEncoder(out, null)
+        val encoder = encoderToUse(avroSchema, out)
         val writer = createDatumWriter(data, avroSchema, useSchemaReflection = useSchemaReflection)
         writer.write(data, encoder)
         encoder.flush()
@@ -70,5 +74,12 @@ class AbstractConfluentKafkaAvroSerializer(avroSchemaEvolution: AvroSchemaEvolut
     val bytes = out.toByteArray
     out.close()
     bytes
+  }
+
+  protected def encoderToUse(schema: Schema, out: OutputStream): Encoder = this.encoderFactory.directBinaryEncoder(out, null)
+
+  protected def writeHeader(data: Any, avroSchema: Schema, schemaId: Int, out: OutputStream): Unit = {
+    out.write(AbstractKafkaSchemaSerDe.MAGIC_BYTE)
+    out.write(ByteBuffer.allocate(AbstractKafkaSchemaSerDe.idSize).putInt(schemaId).array)
   }
 }
