+++ engine/flink/avro-util/src/test/scala/pl/touk/nussknacker/engine/avro/KafkaJsonPayloadSpec.scala
@@ -1,39 +1,30 @@
 package pl.touk.nussknacker.engine.avro
 
-import java.nio.charset.StandardCharsets
-
 import org.apache.avro.Schema
-import org.scalatest.BeforeAndAfter
+import org.apache.kafka.common.serialization.Deserializer
+import org.scalatest.{BeforeAndAfter, FunSuite}
 import pl.touk.nussknacker.engine.api.process.ProcessObjectDependencies
 import pl.touk.nussknacker.engine.avro.KafkaAvroTestProcessConfigCreator.recordingExceptionHandler
+import pl.touk.nussknacker.engine.avro.helpers.{KafkaAvroSpecMixin, SimpleKafkaJsonDeserializer, SimpleKafkaJsonSerializer}
 import pl.touk.nussknacker.engine.avro.schema.PaymentV1
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.ConfluentSchemaRegistryProvider
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.{ConfluentSchemaRegistryClientFactory, MockConfluentSchemaRegistryClientFactory, MockSchemaRegistryClient}
-import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.jsonpayload.{ConfluentJsonPayloadDeserializerFactory, ConfluentJsonPayloadSerializerFactory}
 import pl.touk.nussknacker.engine.avro.schemaregistry.{ExistingSchemaVersion, SchemaRegistryProvider}
-import pl.touk.nussknacker.engine.kafka.KafkaConfig
 import pl.touk.nussknacker.engine.process.compiler.FlinkProcessCompiler
 import pl.touk.nussknacker.engine.process.registrar.FlinkProcessRegistrar
 import pl.touk.nussknacker.engine.testing.LocalModelData
-import pl.touk.nussknacker.engine.util.json.BestEffortJsonEncoder
 
-class KafkaJsonPayloadSpec  extends KafkaAvroSpecMixin with BeforeAndAfter {
+class KafkaJsonPayloadSpec extends FunSuite with KafkaAvroSpecMixin with BeforeAndAfter {
 
   import KafkaAvroIntegrationMockSchemaRegistry._
 
   private lazy val creator: KafkaAvroTestProcessConfigCreator = new KafkaAvroTestProcessConfigCreator {
     override protected def createSchemaRegistryProvider(processObjectDependencies: ProcessObjectDependencies): SchemaRegistryProvider = {
       val clientFactory = new MockConfluentSchemaRegistryClientFactory(schemaRegistryMockClient)
-      val kafkaConfig = KafkaConfig.parseConfig(processObjectDependencies.config)
-      new ConfluentSchemaRegistryProvider(clientFactory,
-        new ConfluentJsonPayloadSerializerFactory(clientFactory),
-        new ConfluentJsonPayloadDeserializerFactory(clientFactory),
-        kafkaConfig, false)
+      ConfluentSchemaRegistryProvider.jsonPayload(clientFactory, processObjectDependencies, formatKey = false)
     }
   }
-
-  private val encoder = BestEffortJsonEncoder(failOnUnkown = false)
-
+  
   protected val paymentSchemas: List[Schema] = List(PaymentV1.schema)
 
   override def schemaRegistryClient: MockSchemaRegistryClient = schemaRegistryMockClient
@@ -56,14 +47,10 @@ class KafkaJsonPayloadSpec  extends KafkaAvroSpecMixin with BeforeAndAfter {
     val sinkParam = SinkAvroParam(topicConfig, ExistingSchemaVersion(1), "#input")
     val process = createAvroProcess(sourceParam, sinkParam)
 
-    runAndVerifyResult(process, topicConfig, PaymentV1.exampleData, encoder.encode(PaymentV1.exampleData))
+    runAndVerifyResult(process, topicConfig, PaymentV1.exampleData, valueSerializer.encoder.encode(PaymentV1.exampleData))
   }
 
-  override protected def serialize(objectTopic: String, obj: Any): Array[Byte] = {
-    encoder.encode(obj).spaces2.getBytes(StandardCharsets.UTF_8)
-  }
+  override protected def prepareValueDeserializer(useSpecificAvroReader: Boolean): Deserializer[Any] = SimpleKafkaJsonDeserializer
 
-  override protected def deserialize(useSpecificAvroReader: Boolean)(objectTopic: String, obj: Array[Byte]): Any = {
-    io.circe.parser.parse(new String(obj, StandardCharsets.UTF_8)).right.get
-  }
+  override protected def valueSerializer: SimpleKafkaJsonSerializer.type = SimpleKafkaJsonSerializer
 }
