+++ engine/flink/avro-util/src/test/scala/pl/touk/nussknacker/engine/avro/schemaregistry/confluent/serialization/ConfluentKafkaAvroDeserializationSpec.scala
@@ -1,35 +1,32 @@
-package pl.touk.nussknacker.engine.avro.serialization
+package pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization
 
 import io.confluent.kafka.schemaregistry.client.{SchemaRegistryClient => CSchemaRegistryClient}
 import org.apache.avro.Schema
 import org.apache.avro.generic.GenericRecord
 import org.apache.kafka.common.errors.SerializationException
 import org.scalatest.Assertion
-import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor4}
+import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor5}
+import pl.touk.nussknacker.engine.avro.RuntimeSchemaData
+import pl.touk.nussknacker.engine.avro.helpers.SchemaRegistryMixin
 import pl.touk.nussknacker.engine.avro.schema.{FullNameV1, PaymentV1, PaymentV2}
 import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.ConfluentUtils
-import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.client.ConfluentSchemaRegistryClientFactory
-import pl.touk.nussknacker.engine.avro.schemaregistry.confluent.serialization.ConfluentKafkaAvroDeserializationSchemaFactory
-import pl.touk.nussknacker.engine.avro.{KafkaAvroSpecMixin, RuntimeSchemaData}
 
-class ConfluentKafkaAvroDeserializationSpec extends KafkaAvroSpecMixin with TableDrivenPropertyChecks with ConfluentKafkaAvroSeDeSpecMixin {
+class ConfluentKafkaAvroDeserializationSpec extends SchemaRegistryMixin with TableDrivenPropertyChecks with ConfluentKafkaAvroSeDeSpecMixin {
 
   import MockSchemaRegistry._
 
   override protected def schemaRegistryClient: CSchemaRegistryClient = schemaRegistryMockClient
 
-  override protected def confluentClientFactory: ConfluentSchemaRegistryClientFactory = factory
-
-  private val confluentDeserializationSchemaFactory = new ConfluentKafkaAvroDeserializationSchemaFactory(factory)
-
   test("should properly deserialize record to avro object with same schema version") {
     val schemas = List(PaymentV1.schema)
     val version = Some(1)
 
-    val table = Table[Boolean, GenericRecord, GenericRecord, String](
-      ("schemaEvolution", "givenObj", "expectedObj", "topic"),
-      (false, PaymentV1.record, PaymentV1.record, "simple.from-record"),
-      (true, PaymentV1.record, PaymentV1.record, "simple.from-subject-version")
+    val table = Table[SchemaRegistryProviderSetup, Boolean, Any, GenericRecord, String](
+      ("setup", "schemaEvolution", "givenObj", "expectedObj", "topic"),
+      (avroSetup, false, PaymentV1.record, PaymentV1.record, "simple.from-record"),
+      (avroSetup, true, PaymentV1.record, PaymentV1.record, "simple.from-subject-version"),
+      (jsonSetup, true, PaymentV1.exampleData, PaymentV1.record, "json.simple.from-subject-version")
+
     )
 
     runDeserializationTest(table, version, schemas)
@@ -39,10 +36,11 @@ class ConfluentKafkaAvroDeserializationSpec extends KafkaAvroSpecMixin with Tabl
     val schemas = List(PaymentV1.schema, PaymentV2.schema)
     val version = Some(2)
 
-    val table = Table[Boolean, GenericRecord, GenericRecord, String](
-      ("schemaEvolution", "givenObj", "expectedObj", "topic"),
-      (false, PaymentV1.record, PaymentV1.record, "forward.from-record"),
-      (true, PaymentV1.record, PaymentV2.record, "forward.from-subject-version")
+    val table = Table[SchemaRegistryProviderSetup, Boolean, Any, GenericRecord, String](
+      ("setup", "schemaEvolution", "givenObj", "expectedObj", "topic"),
+      (avroSetup, false, PaymentV1.record, PaymentV1.record, "forward.from-record"),
+      (avroSetup, true, PaymentV1.record, PaymentV2.record, "forward.from-subject-version"),
+      (jsonSetup, true, PaymentV1.exampleData, PaymentV2.record, "json.forward.from-subject-version")
     )
 
     runDeserializationTest(table, version, schemas)
@@ -52,10 +50,11 @@ class ConfluentKafkaAvroDeserializationSpec extends KafkaAvroSpecMixin with Tabl
     val schemas = List(PaymentV1.schema, PaymentV2.schema)
     val version = None
 
-    val table = Table[Boolean, GenericRecord, GenericRecord, String](
-      ("schemaEvolution", "givenObj", "expectedObj", "topic"),
-      (false, PaymentV1.record, PaymentV1.record, "forward.latest.from-record"),
-      (true, PaymentV1.record, PaymentV2.record, "forward.latest.from-subject-version")
+    val table = Table[SchemaRegistryProviderSetup, Boolean, Any, GenericRecord, String](
+      ("setup", "schemaEvolution", "givenObj", "expectedObj", "topic"),
+      (avroSetup, false, PaymentV1.record, PaymentV1.record, "forward.latest.from-record"),
+      (avroSetup, true, PaymentV1.record, PaymentV2.record, "forward.latest.from-subject-version"),
+      (jsonSetup, true, PaymentV1.exampleData, PaymentV2.record, "json.forward.latest.from-subject-version")
     )
 
     runDeserializationTest(table, version, schemas)
@@ -65,10 +64,11 @@ class ConfluentKafkaAvroDeserializationSpec extends KafkaAvroSpecMixin with Tabl
     val schemas = List(PaymentV1.schema, PaymentV2.schema)
     val version = Some(1)
 
-    val table = Table[Boolean, GenericRecord, GenericRecord, String](
-      ("schemaEvolution", "givenObj", "expectedObj", "topic"),
-      (false, PaymentV2.record, PaymentV2.record, "backward.from-record"),
-      (true, PaymentV2.record, PaymentV1.record, "backward..from-subject-version")
+    val table = Table[SchemaRegistryProviderSetup, Boolean, Any, GenericRecord, String](
+      ("setup", "schemaEvolution", "givenObj", "expectedObj", "topic"),
+      (avroSetup, false, PaymentV2.record, PaymentV2.record, "backward.from-record"),
+      (avroSetup, true, PaymentV2.record, PaymentV1.record, "backward.from-subject-version"),
+      (jsonSetup, true, PaymentV2.exampleData, PaymentV1.record, "json.backward.from-subject-version")
     )
 
     runDeserializationTest(table, version, schemas)
@@ -79,10 +79,11 @@ class ConfluentKafkaAvroDeserializationSpec extends KafkaAvroSpecMixin with Tabl
     val fromRecordTopic = createAndRegisterTopicConfig("wrong.from-record", schemas)
     val fromSubjectVersionTopic = createAndRegisterTopicConfig("wrong.from-subject-version", schemas)
 
+
     pushMessage(FullNameV1.record, fullNameTopic, Some(fromRecordTopic.input))
     pushMessage(FullNameV1.record, fullNameTopic, Some(fromSubjectVersionTopic.input))
 
-    val fromRecordDeserializer = confluentDeserializationSchemaFactory.create(None, kafkaConfig)
+    val fromRecordDeserializer = avroSetup.provider.deserializationSchemaFactory.create(None, kafkaConfig)
 
     consumeAndVerifyMessages(fromRecordDeserializer, fromRecordTopic.input, List(FullNameV1.record))
 
@@ -90,7 +91,7 @@ class ConfluentKafkaAvroDeserializationSpec extends KafkaAvroSpecMixin with Tabl
       val subject = ConfluentUtils.topicSubject(fromSubjectVersionTopic.input, fromSubjectVersionTopic.isKey)
       val schemaId = schemaRegistryClient.getId(subject, ConfluentUtils.convertToAvroSchema(PaymentV1.schema))
       val schemaData = RuntimeSchemaData(PaymentV1.schema, Some(schemaId))
-      confluentDeserializationSchemaFactory.create(Some(schemaData), kafkaConfig)
+      avroSetup.provider.deserializationSchemaFactory.create(Some(schemaData), kafkaConfig)
     }
 
     assertThrows[SerializationException] {
@@ -98,8 +99,9 @@ class ConfluentKafkaAvroDeserializationSpec extends KafkaAvroSpecMixin with Tabl
     }
   }
 
-  private def runDeserializationTest(table: TableFor4[Boolean, GenericRecord, GenericRecord, String], version: Option[Int], schemas: List[Schema]): Assertion =
-    forAll(table) { (schemaEvolution: Boolean, givenObj: GenericRecord, expectedObj: GenericRecord, topic: String) =>
+  private def runDeserializationTest(table: TableFor5[SchemaRegistryProviderSetup, Boolean, Any, GenericRecord, String], version: Option[Int], schemas: List[Schema]): Assertion = {
+
+    forAll(table) { (setup: SchemaRegistryProviderSetup, schemaEvolution: Boolean, givenObj: Any, expectedObj: GenericRecord, topic: String) =>
       val topicConfig = createAndRegisterTopicConfig(topic, schemas)
 
       val schemaDataOpt = if (schemaEvolution) {
@@ -109,10 +111,13 @@ class ConfluentKafkaAvroDeserializationSpec extends KafkaAvroSpecMixin with Tabl
       } else {
         None
       }
-      val deserializer = confluentDeserializationSchemaFactory.create(schemaDataOpt, kafkaConfig)
+      val deserializer = setup.provider.deserializationSchemaFactory.create(schemaDataOpt, kafkaConfig)
 
-      pushMessage(givenObj, topicConfig.input)
+      setup.pushMessage(givenObj, topicConfig.input)
 
-      consumeAndVerifyMessages(deserializer, topicConfig.input, List(expectedObj))
+      setup.consumeAndVerifyMessages(deserializer, topicConfig.input, List(expectedObj))
     }
+
+  }
+
 }
